roleベースの認可処理

+++ src/items/items.service.ts
@@ -1,8 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateItemDto } from './dto/create-item.dto';
 import { Item } from '../entities/item.entity';
 import { ItemRepository } from './item.repository';
 import { ItemStatus } from './item-status.enum';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class ItemsService {
@@ -18,18 +23,27 @@ export class ItemsService {
     return found;
   }
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
-    return await this.itemRepository.createItem(createItemDto);
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
+    return await this.itemRepository.createItem(createItemDto, user);
   }
 
-  async updateStatus(id: string): Promise<void> {
+  async updateStatus(id: string, user: User): Promise<void> {
+    const item = await this.itemRepository.findOne({ id });
+    if (item.userId !== user.id)
+      throw new BadRequestException('出品者でないと更新できません');
+
     await this.itemRepository.update(id, {
       status: ItemStatus.SOLD_OUT,
       updatedAt: new Date().toISOString(),
     });
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.itemRepository.findOne({ id });
+
+    if (item.userId !== user.id)
+      throw new BadRequestException('出品者でないと削除できません');
+
     await this.itemRepository.delete({ id });
   }
 }
